This code adds MD2, MD4 and MD5 file verification
This code allows for the addition of future MD2, MD4 and MD5 file verification.

When the proper code is added into the Paperback block ECC structure a file can be known also by its  (MD2, MD4 and MD5) hashsum and later verified if it recovered but damaged. CRC16_CCITT needs backup!